Create utils folder and move in all utility and types files
