+++ src/containers/User/UserArea.tsx
@@ -1,7 +1,7 @@
 import React, { useEffect } from 'react'
 import { Button, Container, Label, Segment } from 'semantic-ui-react'
 import { useUserState } from './UserState'
-import { useGetUsersQuery, User } from '../../generated/graphql'
+import { useGetUsersQuery, User } from '../../utils/generated/graphql'
 import Loading from '../../components/Loading'
 
 const UserArea: React.FC = () => {
