+++ src/utils/graphql/fragments/addNewUser.fragment.ts
