+++ src/utils/graphql/mutations/createApplication.mutation.ts
