+++ src/utils/graphql/queries/getApplication.query.ts
