+++ src/utils/graphql/queries/getApplications.query.ts
