+++ src/utils/types.ts
@@ -1,0 +1,15 @@
+export {
+    ApplicationPayload,
+    SectionPayload
+}
+
+interface ApplicationPayload {
+    serialNumber: string
+    templateId: number
+    templateName: string
+  }
+  
+interface SectionPayload {
+    applicationId: number
+    templateSections: number[]
+}
